fix recipe viewing / ingredient amount -> f32

+++ api/src/data.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Serialize};
 #[derive(FromQueryResult, Serialize, Deserialize, Debug)]
 pub struct PartialCombinedRecipeIngredient {
     pub name: String,
-    pub amount: i32,
+    pub amount: f32,
     pub description: String,
     pub id: i32,
 }
